Read parameters for point separation from a config file

+++ shapefile/XYZSeparator/Program.cs
@@ -11,16 +11,28 @@ using System.IO;
 namespace XYZSeparator {
 	class Program {
 		public static void Main(string[] args) {
-			string inputFolder = "E:/pointdata/xyz/";
-			string outputFolder = "E:/pointdata/output/";
-			string shapeFileName = "E:/pointdata/shape/shapes.shp";
-			
-			foreach (var file in new DirectoryInfo(outputFolder).GetFiles()) {
-				file.Delete();
+			var config = new ConfigReader("config.ini");
+
+			string inputFolder = config.Get("inputFolder");
+			if (!inputFolder.EndsWith("/") && !inputFolder.EndsWith("\\")) {
+				inputFolder += "/";
+			}
+			string outputFolder = config.Get("outputFolder");
+			if (!outputFolder.EndsWith("/") && !outputFolder.EndsWith("\\")) {
+				outputFolder += "/";
+			}
+			string shapeFilename = config.Get("shapefile");
+			string metadataFilename = config.Get("metadataFile");
+
+			if (config.GetBool("clearOutputFolder")) {
+				Console.WriteLine("Clearing output folder...");
+				foreach (var file in new DirectoryInfo(outputFolder).GetFiles()) {
+					file.Delete();
+				}
 			}
 
-			var shapeHashSet = new ShapeHashSet(100, 2);
-			shapeHashSet.Load(shapeFileName);
+			var shapeHashSet = new ShapeHashSet(100, config.GetDouble("offset"));
+			shapeHashSet.Load(shapeFilename);
 
 			PointSeparator separator = new PointSeparator(shapeHashSet, outputFolder);
 
@@ -33,8 +45,9 @@ namespace XYZSeparator {
 
 			separator.Run();
 
-			// This takes too much memory
-			//Polygon.SaveAggregatedMetadata(outputFolder);
+			Polygon.SaveAggregatedMetadata(metadataFilename);
+
+			Console.WriteLine("Complete.");
 			Console.ReadLine();
 		}
 	}
